Fixed a bug that writes the number of data points in float rather than int.
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@967 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ tests/frontend/CepstrumFileDumper.java
@@ -119,7 +119,7 @@ public class CepstrumFileDumper {
     public void dumpBinary(String outputFile) throws IOException {
 	DataOutputStream outStream = new DataOutputStream
 	    (new FileOutputStream(outputFile));
-	outStream.writeFloat((float) getNumberDataPoints());
+	outStream.writeInt(getNumberDataPoints());
 
 	for (Iterator i = allFeatures.iterator(); i.hasNext();) {
 	    Feature feature = (Feature) i.next();
